Moved all vision files to their own folder
