+++ RobotCode/src/main/java/frc/robot/subsystems/Vision/Camera.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems.Drive;
+package frc.robot.subsystems.Vision;
 
 import java.util.List;
 import java.util.Optional;
@@ -18,8 +18,6 @@ import org.photonvision.simulation.VisionSystemSim;
 import org.photonvision.targeting.PhotonPipelineResult;
 import org.photonvision.targeting.PhotonTrackedTarget;
 
-import edu.wpi.first.apriltag.AprilTagFieldLayout;
-import edu.wpi.first.apriltag.AprilTagFields;
 import edu.wpi.first.math.Matrix;
 import edu.wpi.first.math.VecBuilder;
 import edu.wpi.first.math.geometry.Pose2d;
@@ -37,6 +35,7 @@ import frc.robot.Robot;
 import frc.robot.settings.FieldConstants;
 import frc.robot.settings.VisionConstants;
 import frc.robot.settings.VisionConstants.CamerasConstants;
+import frc.robot.subsystems.Drive.CommandSwerveDrivetrain;
 
 public class Camera extends SubsystemBase {
     // Save the previous result for reference pose strategy
